+++ ml_api/test_data/1.json
@@ -1,0 +1,1 @@
+[{"name": "failure", "confidence": 0.6891447305679321, "box": {"xc": 535.3487373590469, "yc": 474.9534823894501, "w": 252.1463025212288, "h": 138.7461107969284}}, {"name": "failure", "confidence": 0.6743507385253906, "box": {"xc": 263.38869220018387, "yc": 544.919519662857, "w": 276.50628340244293, "h": 185.15091562271118}}, {"name": "failure", "confidence": 0.5817742347717285, "box": {"xc": 589.4390723705292, "yc": 481.2896337509155, "w": 121.88863205909729, "h": 94.02524149417877}}, {"name": "failure", "confidence": 0.47977277636528015, "box": {"xc": 405.2845135331154, "yc": 396.83880066871643, "w": 173.75898277759552, "h": 80.74439334869385}}, {"name": "failure", "confidence": 0.33845505118370056, "box": {"xc": 349.38735204935074, "yc": 382.78335213661194, "w": 164.63774612545967, "h": 82.647167801857}}]
